+++ app/models/payment/Payment.scala
@@ -2,7 +2,7 @@ package models.payment
 
 import play.api.libs.json.Json
 
-case class Payment(id: Int, paymentType: String)
+case class Payment(id: Long, paymentType: String)
 
 object Payment {
   implicit val paymentJSON = Json.format[Payment]
