+++ app/models/user/User.scala
@@ -2,7 +2,7 @@ package models.user
 
 import play.api.libs.json.Json
 
-case class User(id: Int, username: String, password: String)
+case class User(id: Long, username: String, password: String)
 
 object User {
   implicit val userJSON = Json.format[User]
